Flask服务器端/requirements/requirements.txt
absl-py==0.7.1
alembic==1.4.0
APScheduler==3.6.3
asn1crypto==0.24.0
astor==0.8.0
attrs==19.1.0
Automat==0.7.0
blinker==1.4
cachetools==3.1.1
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.8
cssselect==1.0.3
cycler==0.10.0
dominate==2.4.0
Faker==4.0.3
Flask==1.1.1
Flask-APScheduler==1.11.0
Flask-Bootstrap==3.3.7.1
Flask-Email==1.4.4
Flask-HTTPAuth==3.3.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gast==0.2.2
google-api-core==1.13.0
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-translate==1.5.0
google-pasta==0.1.7
googleapis-common-protos==1.6.0
grpcio==1.21.1
h5py==2.9.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.13.2
jsonpath==0.82
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.3.4
Mako==1.1.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mysql-connector-python==8.0.16
newsapi-python==0.2.6
numpy==1.16.4
pandas==0.24.2
parsel==1.5.1
plotly==4.5.0
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.4.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.1
queuelib==1.5.0
requests==2.22.0
retrying==1.3.3
rsa==4.0
scikit-learn==0.21.2
scipy==1.3.0
service-identity==18.1.0
six==1.14.0
sklearn==0.0
SQLAlchemy==1.3.13
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
text-unidecode==1.3
tzlocal==2.0.0
urllib3==1.25.3
visitor==0.1.3
w3lib==1.20.0
Werkzeug==0.16.1
wrapt==1.11.2
WTForms==2.2.1
you-get==0.4.1328
zope.interface==4.6.0
